refactor(frontend): move ACL and role components under security directory
Reorganize security-related components to mirror the route structure.
ACL pages, user pages, role pages, and shared utilities now live under
src/components/pages/security/{acls,users,roles,shared} instead of
the flat pages/acls/ and pages/roles/ directories.

- acls/: ACL create/detail/update pages and form components
- users/: user create/details, edit modals, and user card components
- roles/: role create/detail/update pages, matching users card
- shared/: acl-model (renamed from acl.model), principal-utils,
  operations-badge, acl-details, autocomplete-input
