+++ frontend/src/components/pages/security/acls/host-selector.tsx
@@ -11,10 +11,10 @@
 
 import { useNavigate } from '@tanstack/react-router';
 
-import type { AclDetail } from './acl.model';
-import { Card, CardContent, CardHeader, CardTitle } from '../../redpanda-ui/components/card';
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '../../redpanda-ui/components/table';
-import { Text } from '../../redpanda-ui/components/typography';
+import { Card, CardContent, CardHeader, CardTitle } from '../../../redpanda-ui/components/card';
+import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '../../../redpanda-ui/components/table';
+import { Text } from '../../../redpanda-ui/components/typography';
+import type { AclDetail } from '../shared/acl-model';
 
 type HostSelectorProps = {
   principalName: string;
