+++ frontend/src/components/pages/security/shared/principal-utils.test.ts
