+++ frontend/src/react-query/api/acl.tsx
@@ -41,7 +41,7 @@ import {
   getIdFromCreateACLRequest,
   type Rule,
   type SharedConfig,
-} from '../../components/pages/acls/acl.model';
+} from '../../components/pages/security/shared/acl-model';
 
 /**
  * TODO: Remove once Console v3 is released.
