+++ frontend/tests/test-variant-console/utils/acl-page.ts
@@ -9,7 +9,7 @@ import {
   OperationTypeAllow,
   OperationTypeDeny,
   type Rule,
-} from '../../../src/components/pages/acls/acl.model';
+} from '../../../src/components/pages/security/shared/acl-model';
 
 /**
  * Page Object Model for ACL (Access Control List) pages
